// src/model/Application.ts
export interface Application {
    archiveReference:   string;
    caseId:             string;
    accreditationId:    null;
    accreditationDate:  null;
    accreditationCount: null;
    archivedDate:       string;
    requestor:          string;
    requestorName:      string;
    subject:            string;
    subjectName:        string;
    businessAddress:    null;
    phone:              null;
    email:              null;
    serviceCode:        null;
    languageCode:       null;
    status:             string;
    form:               null;
    attachments:        any;
    consents:           any;
    version:            number;
    updatedDate:        string;
    createdDate:        null;
}